tests(modkit-db): reorganize tests
