+++ libs/modkit-db/tests/odata/sqlite/apply_ext_methods.rs
